+++ CloudHarvestApi/__register__.py
@@ -1,10 +1,9 @@
 """
 The `__register__.py` module is used to register objects that should be placed in the
-`CloudHarvestCorePluginManager.registry.Registry`. Here, they are loaded via the `wsgi.py` -> `api/launch.py`
+`CloudHarvestCorePluginManager.registry.Registry`. Here, they are loaded via the `app.py` -> `api/launch.py`
 startup process. Note that even though these modules do not appear to be used, they are necessary to register both
 Tasks and Blueprints for the Flask application.
 """
 
-# Register definitions
-
-# Register instances
+# Register Blueprints
+from blueprints import *