+++ CloudHarvestApi/cache/exceptions.py
@@ -1,0 +1,6 @@
+from CloudHarvestCoreTasks.exceptions import BaseHarvestException
+
+
+class HarvestCacheException(BaseHarvestException):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)